backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.196
transformers==4.35.2
openai==1.3.6
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
easyocr==1.7.0
